requirements.txt
django <3
gunicorn
djangorestframework
